requirements.txt
Django==4.2.1
django-cors-headers==4.0.0
django-oauth-toolkit==2.2.0
djangorestframework==3.14.0
drf-social-oauth2==2.1.2
mysqlclient==2.1.1
openai==0.27.7
Pillow==9.5.0
python-decouple==3.8
social-auth-app-django==5.2.0
social-auth-core==4.4.2
stripe==5.4.0
